+++ backend/src/main/java/at/fhtw/backend/utils/LocationUtils.java
@@ -3,7 +3,7 @@ package at.fhtw.backend.utils;
 import at.fhtw.backend.model.entities.Location;
 
 public class LocationUtils {
-    public static float[] getCoordinates(Location location) {
-        return new float[] {location.getLng(), location.getLat()};
+    public static double[] getCoordinates(Location location) {
+        return new double[] {location.getLng(), location.getLat()};
     }
 }
